+++ src/components/PlanningHistoric/components/Historic/index.ts
@@ -1,0 +1,1 @@
+export * from './Historic';
